[Fix] Add missing module_id to payment.acquirer data
This commit fixes a bug in the Paybox module's V13 migration.

The payment.acquirer data record was missing the module_id field.
This omission prevented the automated creation of the default journal.

The module_id field has been added to the data record to correctly link
the payment acquirer to its parent module, ensuring the journal is
created as intended and allowing for a successful module installation.

Signed-off-by: Logan Gônet <logan.gonet@horanet.com>